requirements.txt
streamlit
pandas
scikit-learn
mysql-connector-python
nltk
